Add acoustic BSDF
This PR adds an Acoustic BSDF.

The BSDF is equivalent to a BlendBSDF where the first BSDF is a roughconductor and the second BSDF is a diffuse BSDF.

The weighting is done using `scattering`, which is evaluated at the ray's frequency (stored in `si.wavelengths`). If we migrate to the spectral variants, we need to update `eval_scattering` and implement Ratio Control Variates.